Optimize state update and separate filtering logic
